Tmp fixes

+++ sockets.py
@@ -6,8 +6,11 @@ import asyncio
 import json
 import websockets
 from aiohttp import web
+import datetime
+
 # nix-shell -p python311Packages.websockets
 data = dict()
+
 
 async def handler(websocket, path):
     try:
@@ -22,43 +25,85 @@ async def handler(websocket, path):
             except Exception as e:
                 print(e)
     except websockets.exceptions.ConnectionClosed as e:
-        print(e,path)
+        a = datetime.datetime.now()
+        drain_for_module(path)
+        b = datetime.datetime.now()
+        delta = b - a
+        print("time taken to dump: ", path[1:], delta)
     except Exception as e:
         print(e)
 
-def process_fdep_output(k,v):
-    os.makedirs(k[1:].replace(".json",""), exist_ok=True)
-    with open(k[1:],'w') as f:
-        json.dump(v,f,indent=4)
+
+def drain_for_module(path):
+    global data
+    if data.get(path) != None:
+        v = data.get(path)
+        try:
+            process_fdep_output(path, v)
+            del data[path]
+        except Exception as e:
+            print("draining", e)
+
+
+def process_fdep_output(k, v):
+    if not os.path.isfile(k[1:]):
+        os.makedirs(k[1:].replace(".json", ""), exist_ok=True)
+        with open(k[1:], "w") as f:
+            json.dump(v, f, indent=4)
+    else:
+        try:
+            with open(k[1:], "r") as f:
+                alreadyPresentdict = json.load(f)
+                newDict = dict(v, **alreadyPresentdict)
+                print(v.keys(), alreadyPresentdict.keys())
+                with open(k[1:], "w") as ff:
+                    json.dump(newDict, ff, indent=4)
+        except Exception as e:
+            print(e)
+            with open(k[1:], "w") as f:
+                json.dump(v, f, indent=4)
+
 
 async def drain_data(request):
     global data
     with concurrent.futures.ThreadPoolExecutor(max_workers=20) as executor:
-        future_to_file = {executor.submit(process_fdep_output, k,v): (k,v) for (k,v) in data.items()}
+        future_to_file = {
+            executor.submit(process_fdep_output, k, v): (k, v)
+            for (k, v) in data.items()
+        }
         for future in concurrent.futures.as_completed(future_to_file):
             pass
     print(json.dumps(list(data.keys())))
-    exit()
+    exit(0)
+
 
 async def start_websocket_server():
-    async with websockets.serve(handler, "localhost", 8000,ping_interval=None,ping_timeout=None,close_timeout=None,max_queue=1000):
-        print("WebSocket server started on ws://localhost:8000")
+    async with websockets.serve(
+        handler,
+        "localhost",
+        9898,
+        ping_interval=None,
+        ping_timeout=None,
+        close_timeout=None,
+        max_queue=1000,
+    ):
+        print("WebSocket server started on ws://localhost:9898")
         await asyncio.Future()
+
 
 async def start_http_server():
     app = web.Application()
-    app.router.add_get('/drain', drain_data)
+    app.router.add_get("/drain", drain_data)
     runner = web.AppRunner(app)
     await runner.setup()
-    site = web.TCPSite(runner, 'localhost', 8080)
+    site = web.TCPSite(runner, "localhost", 8080)
     await site.start()
     print("HTTP server started on http://localhost:8080")
 
+
 async def main():
-    await asyncio.gather(
-        start_websocket_server(),
-        start_http_server()
-    )
+    await asyncio.gather(start_websocket_server(), start_http_server())
+
 
 if __name__ == "__main__":
     asyncio.run(main())
